Rebase changes + fixing test
