+++ core/src/main/scala/kafka/security/auth/Acl.scala
@@ -19,11 +19,6 @@ package kafka.security.auth
 
 import kafka.security.authorizer.AclEntry
 import org.apache.kafka.common.resource.ResourcePattern
-import java.nio.charset.StandardCharsets
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import kafka.utils.Json
-import kafka.utils.json.JsonValue
 import org.apache.kafka.common.security.auth.KafkaPrincipal
 
 @deprecated("Use org.apache.kafka.common.acl.AclBinding", "Since 2.5")
@@ -56,23 +51,6 @@ object Acl {
     AclEntry.toJsonCompatibleMap(acls.map(acl =>
       AclEntry(acl.principal, acl.permissionType.toJava, acl.host, acl.operation.toJava)
     ))
-  }
-
-  /**
-    * Parse a JSON string into a JsonValue if possible. `None` is returned if `input` is not valid JSON. This method is currently used
-    * to read the already stored invalid ACLs JSON which was persisted using older versions of Kafka (prior to Kafka 1.1.0). KAFKA-6319
-    */
-  private def parseBytesWithAclFallback(input: Array[Byte]): Option[JsonValue] = {
-    // Before 1.0.1, Json#encode did not escape backslash or any other special characters. SSL principals
-    // stored in ACLs may contain backslash as an escape char, making the JSON generated in earlier versions invalid.
-    // Escape backslash and retry to handle these strings which may have been persisted in ZK.
-    // Note that this does not handle all special characters (e.g. non-escaped double quotes are not supported)
-    Json.tryParseBytes(input) match {
-      case Left(e) =>
-        val escapedInput = new String(input, StandardCharsets.UTF_8).replaceAll("\\\\", "\\\\\\\\")
-        Json.parseFull(escapedInput)
-      case Right(v) => Some(v)
-    }
   }
 }
