+++ core/src/main/scala/kafka/security/authorizer/AclEntry.scala
@@ -17,9 +17,12 @@
 
 package kafka.security.authorizer
 
+import java.nio.charset.StandardCharsets
+
 import kafka.utils.Json
+import kafka.utils.json.JsonValue
 import org.apache.kafka.common.acl.{AccessControlEntry, AclOperation, AclPermissionType}
-import org.apache.kafka.common.acl.AclOperation.{READ, WRITE, CREATE, DESCRIBE, DELETE, ALTER, DESCRIBE_CONFIGS, ALTER_CONFIGS, CLUSTER_ACTION, IDEMPOTENT_WRITE}
+import org.apache.kafka.common.acl.AclOperation.{ALTER, ALTER_CONFIGS, CLUSTER_ACTION, CREATE, DELETE, DESCRIBE, DESCRIBE_CONFIGS, IDEMPOTENT_WRITE, READ, WRITE}
 import org.apache.kafka.common.protocol.Errors
 import org.apache.kafka.common.resource.{ResourcePattern, ResourceType}
 import org.apache.kafka.common.security.auth.KafkaPrincipal
@@ -79,7 +82,7 @@ object AclEntry {
     if (bytes == null || bytes.isEmpty)
       return collection.immutable.Set.empty[AclEntry]
 
-    Json.parseBytes(bytes).map(_.asJsonObject).map { js =>
+    parseBytesWithAclFallback(bytes).map(_.asJsonObject).map { js =>
       //the acl json version.
       require(js(VersionKey).to[Int] == CurrentVersion)
       js(AclsKey).asJsonArray.iterator.map(_.asJsonObject).map { itemJs =>
@@ -90,6 +93,23 @@ object AclEntry {
         AclEntry(principal, permissionType, host, operation)
       }.toSet
     }.getOrElse(Set.empty)
+  }
+
+  /**
+   * Parse a JSON string into a JsonValue if possible. `None` is returned if `input` is not valid JSON. This method is currently used
+   * to read the already stored invalid ACLs JSON which was persisted using older versions of Kafka (prior to Kafka 1.1.0). KAFKA-6319
+   */
+  private def parseBytesWithAclFallback(input: Array[Byte]): Option[JsonValue] = {
+    // Before 1.0.1, Json#encode did not escape backslash or any other special characters. SSL principals
+    // stored in ACLs may contain backslash as an escape char, making the JSON generated in earlier versions invalid.
+    // Escape backslash and retry to handle these strings which may have been persisted in ZK.
+    // Note that this does not handle all special characters (e.g. non-escaped double quotes are not supported)
+    Json.tryParseBytes(input) match {
+      case Left(e) =>
+        val escapedInput = new String(input, StandardCharsets.UTF_8).replaceAll("\\\\", "\\\\\\\\")
+        Json.parseFull(escapedInput)
+      case Right(v) => Some(v)
+    }
   }
 
   def toJsonCompatibleMap(acls: Set[AclEntry]): Map[String, Any] = {
